+++ common/src/lib.rs
@@ -4,6 +4,6 @@ mod net;
 mod service_method;
 
 pub use crate::connection::{ConnectionTrait, ReadonlyConnection};
-pub use crate::message::{EncodableMessage, NetMessage};
+pub use crate::message::{DecodableMessage, EncodableMessage, ReceivableMessage, SendableMessage};
 pub use crate::service_method::ServiceMethodRequest;
 pub use net::{JobId, NetMessageHeader, RawSteamId};