+++ src/connection/unauthenticated.rs
@@ -14,7 +14,9 @@ use futures_util::Stream;
 use futures_util::{FutureExt, Sink};
 use serde::Deserialize;
 use std::future::Future;
-use steam_vent_common::{NetMessage, NetMessageHeader, ReadonlyConnection, ServiceMethodRequest};
+use steam_vent_common::{
+    NetMessageHeader, ReadonlyConnection, ReceivableMessage, ServiceMethodRequest,
+};
 use steam_vent_proto_steam::enums_clientserver::EMsg;
 use steamid_ng::{AccountType, SteamID};
 use thiserror::Error;
@@ -211,7 +213,7 @@ impl ReadonlyConnection for UnAuthenticatedConnection {
             .map(|raw| raw.into_notification())
     }
 
-    fn one_with_header<T: NetMessage + 'static>(
+    fn one_with_header<T: ReceivableMessage + 'static>(
         &self,
     ) -> impl Future<Output = Result<(NetMessageHeader, T)>> + 'static {
         // async block instead of async fn, so we don't have to tie the lifetime of the returned future
@@ -223,12 +225,12 @@ impl ReadonlyConnection for UnAuthenticatedConnection {
         }
     }
 
-    fn one<T: NetMessage + 'static>(&self) -> impl Future<Output = Result<T>> + 'static {
+    fn one<T: ReceivableMessage + 'static>(&self) -> impl Future<Output = Result<T>> + 'static {
         self.one_with_header::<T>()
             .map(|res| res.map(|(_, msg)| msg))
     }
 
-    fn on_with_header<T: NetMessage + 'static>(
+    fn on_with_header<T: ReceivableMessage + 'static>(
         &self,
     ) -> impl Stream<Item = Result<(NetMessageHeader, T)>> + 'static {
         BroadcastStream::new(self.0.filter.on_kind(T::KIND)).map(|raw| {
@@ -237,7 +239,7 @@ impl ReadonlyConnection for UnAuthenticatedConnection {
         })
     }
 
-    fn on<T: NetMessage + 'static>(&self) -> impl Stream<Item = Result<T>> + 'static {
+    fn on<T: ReceivableMessage + 'static>(&self) -> impl Stream<Item = Result<T>> + 'static {
         self.on_with_header::<T>()
             .map(|res| res.map(|(_, msg)| msg))
     }