+++ src/lib.rs
@@ -15,5 +15,6 @@ pub use net::{NetworkError, RawNetMessage};
 pub use serverlist::{DiscoverOptions, ServerDiscoveryError, ServerList};
 pub use session::{ConnectionError, LoginError};
 pub use steam_vent_common::{
-    ConnectionTrait, NetMessage, NetMessageHeader, ReadonlyConnection, ServiceMethodRequest,
+    ConnectionTrait, NetMessageHeader, ReadonlyConnection, ReceivableMessage, SendableMessage,
+    ServiceMethodRequest,
 };